// Movies/types.ts
// export type IMovie = {
//     adult: boolean
//     backdrop_path: string
//     genre_ids: Array<number>
//     id: string
//     original_language: string
//     original_title: string
//     overview: string
//     popularity: number
//     poster_path: string
//     release_date: string
//     title: string
//     video: boolean
//     vote_average: number
//     vote_count: number
// }


export type IMovies = {
    poster_path: any
    adult: boolean
    overview: string
    release_date: string
    genre_ids: number[]
    id: string
    original_title: string
    original_language: string
    title: string
    backdrop_path: any
    popularity: number
    vote_count: number
    video: boolean
    vote_average: number
}

export interface IMovie {
    adult:                 boolean;
    backdrop_path:         string;
    belongs_to_collection: BelongsToCollection;
    budget:                number;
    genres:                Genre[];
    homepage:              string;
    id:                    number;
    imdb_id:               string;
    original_language:     string;
    original_title:        string;
    overview:              string;
    popularity:            number;
    poster_path:           string;
    production_companies:  ProductionCompany[];
    production_countries:  ProductionCountry[];
    release_date:          Date;
    revenue:               number;
    runtime:               number;
    spoken_languages:      SpokenLanguage[];
    status:                string;
    tagline:               string;
    title:                 string;
    video:                 boolean;
    vote_average:          number;
    vote_count:            number;
}

export interface BelongsToCollection {
    id:            number;
    name:          string;
    poster_path:   string;
    backdrop_path: string;
}

export interface Genre {
    id:   number;
    name: string;
}

export interface ProductionCompany {
    id:             number;
    logo_path:      string;
    name:           string;
    origin_country: string;
}

export interface ProductionCountry {
    iso_3166_1: string;
    name:       string;
}

export interface SpokenLanguage {
    english_name: string;
    iso_639_1:    string;
    name:         string;
}